+++ ast/src/list.rs
@@ -19,7 +19,7 @@ pub enum ListItemTypeMeta<'arena> {
 }
 
 impl<'arena> ListItem<'arena> {
-  pub const fn loc_start(&self) -> usize {
+  pub const fn loc_start(&self) -> u32 {
     self.marker_src.loc.start
   }
 
@@ -31,7 +31,7 @@ impl<'arena> ListItem<'arena> {
       .or_else(|| self.principle.last_loc())
   }
 
-  pub fn last_loc_end(&self) -> Option<usize> {
+  pub fn last_loc_end(&self) -> Option<u32> {
     self.last_loc().map(|loc| loc.end)
   }
