+++ parser/src/token.rs
@@ -98,11 +98,11 @@ impl<'arena> Token<'arena> {
     }
   }
 
-  pub fn drop_leading_bytes(&mut self, n: usize) {
+  pub fn drop_leading_bytes(&mut self, n: u32) {
     if n == 0 {
       return;
     }
-    debug_assert!(n <= self.lexeme.len());
+    debug_assert!(n as usize <= self.lexeme.len());
     self.kind = TokenKind::Word;
     let mut removed = 0;
     loop {
@@ -110,7 +110,7 @@ impl<'arena> Token<'arena> {
       let mut buf = [0; 4];
       let bytes = char.encode_utf8(&mut buf).as_bytes();
       removed += bytes.len();
-      match removed.cmp(&n) {
+      match removed.cmp(&(n as usize)) {
         Ordering::Less => continue,
         Ordering::Equal => break,
         Ordering::Greater => panic!("Token::drop_leading_bytes() mid-char boundary"),
